requirements.txt
# Comparador de Imágenes con ML - Dependencias Python

# Framework web y servidor
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0

# Procesamiento de imágenes
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.21.0

# Análisis numérico y machine learning
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
joblib==1.3.2

# Utilidades adicionales
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# Logging y debugging
colorlog==6.7.0

# Procesamiento de imágenes avanzado (opcional)
imageio==2.31.5
matplotlib==3.7.2

# Para mejor rendimiento (opcional)
numba==0.58.1

# Google BigQuery y dependencias relacionadas
google-cloud-bigquery==3.13.0
pandas-gbq==0.22.0
pandas==2.1.3
pyarrow==14.0.1

# Autenticación de Google Cloud
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Utilidades adicionales para BigQuery
db-dtypes==1.1.1 